+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1306,7 +1306,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T08:57:11.759Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T09:48:43.411Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1358,7 +1358,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T08:57:11.759Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T09:48:43.411Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1410,7 +1410,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T08:57:11.759Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-10-24T09:48:43.411Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1484,5 +1484,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-24T08:57:01.761Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-24T09:48:33.978Z")
     r"""The date and time the resource was last updated"""