+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8APZBC7QXHMNS2ZB6PBZ4DS"
+        "01K8ASXPHZCQB4CFV5RWA4P2RA"
     )
     r"""The id of the resource"""