+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8APZBBMXQVX19YA4XADXDTH"
+        "01K8ASXPHX6PYTGSM6KJ4JY56Z"
     )
     r"""The id of the resource"""