+++ packages/orq-rc/RELEASES.md
@@ -3589,3 +3589,13 @@ Based on:
 - [python v3.14.0-rc.54] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.54] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.54 - packages/orq-rc
+
+## 2025-10-24 09:50:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.1 (2.730.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.55] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.55] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.55 - packages/orq-rc